+ Use MDX to write documents.
