+++ babel.config.json
@@ -1,0 +1,5 @@
+{
+  "plugins": [
+    "@vue/babel-plugin-jsx"
+  ]
+}
