+++ doc/index.tsx
@@ -1,4 +1,6 @@
+import hljs from 'highlight.js'
 import { defineComponent, createApp } from 'vue'
+import 'highlight.js/styles/rainbow.css'
 
 import 'material-icons/iconfont/material-icons.css'
 import './style/index.styl'
@@ -21,4 +23,10 @@ const Layout = defineComponent({
 const app = createApp(Layout)
 app.use(Lancet)
 app.use(router)
+app.mixin({
+  mounted: () => {
+    hljs.highlightAll()
+  }
+})
+
 app.mount('#app')
