+++ doc/style/index.styl
@@ -1,6 +1,4 @@
-pre
-  margin: 30px 0
-  display: block
-  background: #efefef
-  padding: 10px
-  overflow: auto
+code
+  padding: 5px
+  background-color: #ffe6ea
+  color: #f74b69
