+++ doc/views/home/index.mdx
@@ -1,0 +1,9 @@
+import { LancetLogo } from '../../components/lancet-logo'
+
+<LancetLogo width={300} />
+
+# Lancet
+
+> (Material Design + Vue) * 3
+
+Yet another Material Design like Vue 3 UI Components library.
