--- doc/views/showcase.dynamic-color.module.styl
@@ -1,7 +1,0 @@
-.dynamic-color-showcase
-  .button-list
-    display: flex
-    flex-wrap: wrap
-
-    @media screen and (max-width: $pcBoundary)
-      flex-direction: column
