+++ doc/views/tone/components/select-image.module.styl
@@ -1,0 +1,6 @@
+.select-image
+  display: flex
+  flex-wrap: wrap
+
+  @media screen and (max-width: $pcBoundary)
+    flex-direction: column
