+++ doc/views/tone/components/theme-list.module.styl
@@ -1,0 +1,6 @@
+.theme-list
+  display: flex
+  flex-wrap: wrap
+
+  @media screen and (max-width: $pcBoundary)
+    flex-direction: column
