+++ doc/views/tone/index.tsx
@@ -1,0 +1,20 @@
+import { defineComponent } from 'vue'
+
+import { LctCard, LctCardContent } from '../../../lib'
+
+import Content from './index.mdx'
+
+const DynamicColor = defineComponent({
+  setup () {
+    return () => (
+      <div>
+        <Content />
+
+      </div>
+    )
+  }
+})
+
+export {
+  DynamicColor
+}
