+++ doc/views/typography/index.tsx
@@ -1,0 +1,14 @@
+import { defineComponent } from 'vue'
+import Content from './index.mdx'
+
+const Typography = defineComponent({
+  setup () {
+    return () => (
+      <Content />
+    )
+  }
+})
+
+export {
+  Typography
+}
