+++ package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc --noEmit && vite build",
-    "serve": "vite preview",
-    "prepublish": "npm run build"
+    "serve": "vite preview"
   },
   "files": [
     "dist/"
@@ -14,41 +13,46 @@
   "main": "./dist/index.umd.js",
   "module": "./dist/index.es.js",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "dependencies": {
     "color": "^4.0.1",
     "dayjs": "^1.10.7",
     "vue": "^3.2.22"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.16.0",
-    "@babel/plugin-proposal-decorators": "^7.16.4",
-    "@babel/preset-env": "^7.16.4",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "@types/color": "^3.0.2",
-    "@types/node": "^16.11.10",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "@vitejs/plugin-legacy": "^1.6.3",
-    "@vitejs/plugin-vue": "^1.10.0",
-    "@vitejs/plugin-vue-jsx": "^1.3.0",
-    "@vue/compiler-sfc": "^3.2.22",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/plugin-proposal-decorators": "^7.17.2",
+    "@babel/plugin-syntax-jsx": "^7.16.7",
+    "@babel/preset-env": "^7.16.11",
+    "@mdx-js/rollup": "^2.0.0",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-typescript": "^8.3.1",
+    "@types/color": "^3.0.3",
+    "@types/node": "^16.11.26",
+    "@typescript-eslint/eslint-plugin": "^5.14.0",
+    "@typescript-eslint/parser": "^5.14.0",
+    "@vitejs/plugin-legacy": "^1.7.1",
+    "@vitejs/plugin-vue": "^2.2.4",
+    "@vitejs/plugin-vue-jsx": "^1.3.8",
+    "@vue/babel-plugin-jsx": "^1.1.1",
+    "@vue/compiler-sfc": "^3.2.31",
     "babel-plugin-transform-typescript-metadata": "^0.3.2",
     "color.js": "^1.2.0",
     "cross-env": "^7.0.3",
     "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-vue": "^7.20.0",
-    "material-icons": "^1.10.3",
+    "eslint-plugin-promise": "^5.2.0",
+    "eslint-plugin-vue": "^8.5.0",
+    "highlight.js": "^11.5.0",
+    "material-icons": "^1.10.8",
     "reflect-metadata": "^0.1.13",
     "stylus": "^0.55.0",
-    "typescript": "^4.5.2",
-    "vite": "^2.6.14",
-    "vite-plugin-eslint": "^1.3.0",
+    "typescript": "^4.6.2",
+    "vite": "^2.8.6",
     "vite-svg-loader": "^3.1.0",
     "vue-router": "^4.0.12",
-    "vue-tsc": "^0.3.0"
+    "vue-tsc": "^0.32.1"
   }
 }
