+++ vite.config.ts
@@ -1,9 +1,12 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+
 import path from 'path'
+import mdx from '@mdx-js/rollup'
+import { babel } from '@rollup/plugin-babel'
 import typescript from '@rollup/plugin-typescript'
 import vue from '@vitejs/plugin-vue'
 import vueJsx from '@vitejs/plugin-vue-jsx'
 import { defineConfig } from 'vite'
-import eslintPlugin from 'vite-plugin-eslint'
 import svgLoader from 'vite-svg-loader'
 
 // https://vitejs.dev/config/
@@ -12,6 +15,7 @@ export default defineConfig({
 
   plugins: [
     vue(),
+
     vueJsx({
       babelPlugins: [
         'babel-plugin-transform-typescript-metadata',
@@ -19,9 +23,20 @@ export default defineConfig({
         ['@babel/plugin-proposal-class-properties', { loose: true }]
       ]
     }),
-    eslintPlugin({
-      cache: false
+
+    // @ts-ignore
+    mdx({
+      jsx: true
     }),
+
+    // This babel is for MDX.
+    // It transforms the JSX which is generated by MDX Plugin into JS.
+    // @ts-ignore
+    babel({
+      extensions: ['.md', '.mdx'],
+      babelHelpers: 'bundled'
+    }),
+
     svgLoader()
   ],
 
@@ -69,7 +84,7 @@ export default defineConfig({
           declarationDir: path.resolve(__dirname, './dist'),
           exclude: path.resolve(__dirname, './node_modules/**'),
           allowSyntheticDefaultImports: true
-        })
+        }) as any
       ]
     }
   }
